requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.13.2
python-dotenv==1.1.1
pydantic==2.11.7
python-multipart==0.0.20
dnspython==2.7.0